requirements.txt
# OpenStreetMap Urban Growth Analysis
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.13.0
shapely>=2.0.0

# OSM data handling
osmium==3.6.0
requests==2.31.0
overpy==0.7
pyproj>=3.4.0

# Geospatial analysis
folium==0.15.0
contextily==1.4.0
rasterio==1.3.9
pyproj==3.6.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.28.2
streamlit-folium==0.15.1

# Analysis and ML
scikit-learn==1.3.2
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1

# Development
jupyter==1.0.0
black==23.11.0
flake8==6.1.0
pytest==7.4.3
